Add support for "Get" and "List" parameters
Some of the "Get" and "List" operations of the API support parameters.
For example, the operation to list virtual machines supports the
"search" parameter. This patch adds support for this kind of parameters,
so that this, for example, will work:

  # Create the connection:
  connection = Ovirt::SDK::V4Connection.new({
    :url => 'https://engine.example.com/ovirt-engine/api',
    :username => 'admin@internal',
    :password => '...',
    :ca_file => 'ca.pem',
  })

  # Get the VMs service:
  vms_service = connection.system.vms

  # Search for VMs with a given name and print their id:
  vms = vms_service.list({:search => 'name=myvm'})
  vms.each do |vm|
    puts vm.id
  end

  # Close the connection:
  connection.close

The patch also improves the tests so that this support is better tested.

Change-Id: I0f1fe53af1f81a29c72c4a5ab99f88d4ceda3e1b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>