Use "List" instead of "Array"
Currently when a list of objects is received from the server it is
converted to an instance of the Ruby "Array" class. This patch
introduces a new "List" class, that extends the Ruby "Array" class,
currently with no methods. List of objects returned from the SDK will be
of this type. Later patches will take advantage of this "List" class,
for example to add methods to support following links.

The patch also renames the base "Type" class to "Struct", as that better
reflects the difference between struct and list types.

Change-Id: I1cbcd1eab8bd8b95d7d4a0e0537ccd8a47e81f64
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>