Add path locator methods
This patch adds a "service" method to the connection and to all the
service implementations. This method receives a string parameter
containing the path to a service, for example "vms/123/disks" and
returns a reference to the corresponding service implementation.

The new method is intended to be part of the implementation of link
following that will be added by later patches.

Change-Id: Id16f8b320b27ea1ad347dc116d0f2653b3f72d57
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>