Add "is_link" attribute
This patch adds a new "is_link" attribute to struct and list types. This
attribute indicates if the object is used as a link or as a normal
object. When an object is used as a link the server will only populate
the identifier and "href" attributes. For example, in the specification
of the virtual machine type the "cluster" attribute is defined as a
link:

  @Type
  interface VmBase extends Identified {
    ...
    @Link Cluster cluster();
    ...
  }

This means that the server won't return the complete entity, but just
the identifier and the "href" attributes:

  <vm ...>
    ...
    <cluster id="123" href=".../clusters/123"/>
    ...
  </vm>

The Ruby SDK will mark these uses of the objects, using the "is_link"
flag, so that the users of the SDK can decide, based on the value of
this flag, if furhter loading is needed:

  # Get the VM:
  vm = ...

  # Check if the cluster is a link, in which case we need to
  # load its complete representation explicitly:
  if vm.cluster.is_link?
    vm.cluster = clusters_service.get({:id => vm.cluster.id})
  end

Change-Id: Ifac0d857942e8ca32a2fa5fb24d845a79c4380a7
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>