Add request and response classes
This patch adds two new "Request" and "Response" classes. These classes
are intended to carry the details of HTTP requests and responses. With
them the "Connection.send" method can return a response, containing the
HTTP response code, message and body, so that other parts of the SDK
will in the future be able to inspect them and act accordingly.

Change-Id: I1801639bea2e419f956133e070b7f22dee24985d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>