Add "_service" to all locator methods
Currently the locator methods are named like in the specification of the
API. For example, the method used to find the virtual machines service
is named "vms". This patch changes those methods so that they will have
a "_service" suffix, making it clearer that they return references to
services, not objects.

Change-Id: If3dacbd895fda86b9672643bb045fd23cf2203a0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>