Reduce the number of files
Currently the SDK has one file for each class, which makes for a total
of hundreds of files. This patch changes the code generator so that it
generates related classes in the same files. For example, instead of
having a file for each type, there is now one single "types.rb" file
containing all or them.

Change-Id: Iaa6d87cd34ce1bff3cf2803ead6af7bdc4d1cfe9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>