Add builtin types to the model
Currently the SDK contains types and reader written manually for the
"Fault" and "Action" builtin types. This patch changs the code generator
so that it adds the definitions of these types to the model before
generating the code. As a result the types and readers are also
generated automatically, so there is no need to write them manually.

Change-Id: Id2a08df2f12aa29a4f412443275e8d2667f1a13e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>