Add support for debug log file
Currently when the debug mode is enabled the SDK writes the debug output
to stdout. This isn't convenient for applications that need to write
other things to stdout. This patch adds a new "log" parameter to the
constructor of the connection. This parameter can contain a file name or
an IO object. If it contains a file name the file will be created if
needed, and the debug output will be added to the end of the file. If it
contains an IO object then the debug output will be written directly to
that IO object.

Change-Id: I80e61eb6585ba9f31e95c41ac99615a0681ba2e9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>