Remove the connection attribute from objects
Currently each object created by the SDK contains a "connection"
attribute that is used to implement the "follow_link" method. This is
problematic, because when the connection is closed there will be many
objects that have a non nil but useless connection. To avoid this
problem this patch removes this "connection" attribute and moves the
"follow_link" method to the "Connection" class. So instead of following
a link with this code:

  vm = ...
  cluster = vm.cluster.follow_link

It is now necessary to explicitly use the connection:

  vm = ...
  cluster = connection.follow_link(vm.cluster)

Change-Id: I29202acf545121a18acb9b1c493dc6b613b4418e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>