Support strings as parameters of XML readers and writers
Currently the constructor XML reader and writer classes require IO
objects. But may use cases require passing a string, so there are
multiple places where a "StringIO" object has to be created in order to
satisfy the contracts of the XML readers and writers. To simplify those
use cases this patch improves the XML readers and writers so that they
support also strings. If an IO object isn't provided, then the XML
readers and writers will create "StringIO" objects and use them
internally.

Change-Id: I93e3e2078002a8ad387e0f3ffacb532b586eb52a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>