Use only query parameters
Currently the SDK has to send parameters as both query and matrix
parameters, as the server some times requires query and some times
matrix. But the server has recently been fixed to support both, for all
paramers, so the SDK no longer needs to send both. This patch changes
the SDK so that it only uses query parameters.

Change-Id: Ib17aaaab7b5660619c2ac69ee2f5c737e61d9ce3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>