Use "Action" type
Currently when the SDK needs to perform an action, like starting a
virtual machine, the corresponding service class generates directly the
"<action>...</action>" XML text, and it doesn't extract the output
parameters of some actions. This patch changes it so that instead of
generating directly the XML text it creates an instance of the "Action"
struct, populates it, and then generates the XML text using the
"ActionWriter" class. In also modifies the SDK so that it takes the
returned action and extracts the output parameters.

Change-Id: If2e254cf2b89baa91aaa13a058274a0c0c6b02a0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>