Add support for following links to lists
Currently the "follow_link" method of the "Conection" class only
supports links to single objects, but not links to list of objects. This
patch changes the SDK so that it will also support this kind of links.
For example, when a VM is retrieved it will include links to its NICs:

  <vm id="123" href="/vms/123">
    <link rel="nics" href="/vms/123/nics"/>
    ...
  </vm>

The SDK will process this "link" element and store it in the "vm.nics"
property of the created object. Later, the user will be able to use the
"follow_link" method to retrieve that list:

  vm = ...
  nics = connection.follow_link(vm.nics)
  nics.each do |nic|
    puts "nic: #{nic.name}"
  end

Change-Id: I33adfd45183838f81873ab70fade92c8d89e9c44
Related-To: https://bugzilla.redhat.com/1349857
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit fea6ddb71cf6c1756fe3e84c008bd43915bb05a5)