Clear the POST body before sending requests
Currently the SDK doesn't clear the cURL POST body before sending
requests, as a result the body of a previous POST request can interfere
with a current GET request, causing an error in the server side. This
patch changes the SDK so that the POST body is cleared before every
request.

Change-Id: I9df5a89dec28f4f252dc99fcc774ede017707d30
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>