Handle correctly error responses without body
Currently when the SDK receives a response with an error code but
without a body it generates an incorrect exception because it tries to
create an XML reader that wasn't actually created. To avoid that this
patch changes the SDK so that it checks if the response body is empty,
and in that case it doesn't try to create or close the XML reader.

Change-Id: If898a44b964bce3944e008dc8e10a2866f180452
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>