Add support for Mac OS X
Mac OS X includes a version of "libxml2" that doesn't include a
"pkg-config" configuration file. As a result buliding the gem fails
because it can't find the include files. To avoid that issue this patch
changes the build scripts so that they use the "xml2-config" utility
that is provided by all versions of "libxml", including the one
available in Mac OS X.

In addition the Xcode compiler is a bit more strict and complains about
multiple definitions of some varaibles, this patch fixes that as well.

Change-Id: If18e4373f32a774f33157f52089189341f6a9ffc
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit e8cd7cd9a1c88ec72683fa4430b149a067f30d14)