Bump RPM release number
In previous versions of the SDK the RPM release number used to be
extracted from the version qualifier. For example, if when the version
qualifier was "alpha13" the RPM release was "0.13.alpha13". This isn't
correct, as it complicates bumping the RPM release number when
modifications to the .spec are needed without changes to the SDK. The
ideal solution to this would be to have the .spec files in a different
repository, and update them manually. As that isn't possible the
automation scripts have been changed to automatically generate the RPM
release number, but they don't do this icorrect handling of the release
number, instead they take it directly from the .spec file templates. A
side effect of this is that the RPMs generated by the automation scripts
since the change was introduced look older than previous ones. For
example, the RPM for version "alpha19" looks newer than the RPM for
version "alpha13":

  0.1.alpha19 < 0.13.alpha13

To avoid this issue this patch changes the .spec file templates so that
the release number is greater than the last RPM generated using the
previous approach.

This change will we reverted when there is a change to the SDK major,
minor or fix number. For example, when 4.0.1 is released, the RPM
release number will go back to 0.1.

Change-Id: Iff8ae2e6be1268c01f1a1deb033d9722b91a4cdb
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>