Add generic reader
Currently the SDK doesn't provide any mechanism to read an XML object
without explicitly specifying its type. For example, to read a virtual
machine the only method available is "VmReader.read_one". But there are
situations where the SDK can't know in advance what type of object is
going to be received from the server, for example "action" or "fault".
To support that, this patch adds a new "Reader.read" method that decides
the type of object based on the name of the first tag, and invokes the
corresponding specific reader to read it.

Change-Id: I3b8c9a4c1118e7b261a589e8185861c9acf36921
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>