Add generic writer
Currently the SDK doesn't have any mechanism to generate an XML document
from an object whose type isn't known. This patch adds a new
'Writer.write' method that decides what is the specific writer to use
based on the type of the object, so that the caller doesn't need to
explicitly select it.

Change-Id: I0f6f8af923253707e55b98f04f3407105dc529b4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>