Accept 'action' and 'fault' from actions
Currently the SDK expects that the server to return 'action' XML
elements when actions succeed, and 'fault' when they fail. But the
server may send 'action' with a nested 'fault' when something fails.
This patch modifies the SDK so that it also accepts that.

Change-Id: I9fd210df7aff836421d5c9c3c13652fbd9f390e1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 65a779295c65065749177a9243d1155191d958ee)