Fix reading of lists of elements
Currently the method that read lists of elements doesn't handle
correctly empty elements without close tags. For example, when reading
the following cluster:

  <cluster>
    ...
    <required_rng_sources/>
    <switch_type>legacy</switch_type>
  </cluster>

The 'legacy' value of the 'switch_type' element is interpreted as one of
the values of the previous 'required_rng_sources' list. This patch fixes
the method and introduces two tests to verify that it works correctly.

Change-Id: Id1495e873642aadd09ba3dd578077346f17e873e
Bug-Url: https://bugzilla.redhat.com/1373436
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>