Use XML schema naming exceptions
Currently the SDK uses incorrect XML tag names for some types, in
particular for the 'OpenStackImageProvider' type. For historical reasons
the tag name corresponding to this type is 'openstack_image_provider',
without a dash between the 'open' and 'stack' words. These naming
exceptions are managed by the 'SchemaNames' class of the metamodel, but
we aren't using it correctly.  This patch fixes that error.

Change-Id: Iedc3d7e7bd542ef3d304530588958d591675d9e0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>