Empty list is not written to XML
When the user wants to send an empty list to the server, it is not
currently possible. For example when the user wants to create non VM
network he has to send an empty usages element. But if he sets the
"Network.usages" attribute to empty list, the empty usages element isn't
sent.

This patch fixes that.

Change-Id: Iab4f0af3701b4b812800adc7cc527c4506b1a4ea
Signed-off-by: Ondra Machacek <omachace@redhat.com>
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>