Use test log files for authenticator
The authenticator used by the tests server writes information messages
directly to the standard output, which interferes with the normal
"rspec" output. To avoid that this patch configures the authenticator so
that it writes the messages to the "server.log" file.

Change-Id: I194f2fbc44115068f69cdbdb06f91573474e4ed3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 914e240ca7eccb90901b27a4f9dcda74a235b847)