Remove 'curb' dependency
Currently the SDK requires the 'curb' gem, but this gem isn't available
in some of the target distributions, in particular it isn't available in
Red Hat Enterprise Linux 7. In addition 'curb' doesn't support some
features that we want to use in the future, like streaming of results.
To address these issues this patch replaces the code that uses 'curb'
with code that calls directly 'libcurl'.

Change-Id: Ia3a894fe3d622c92d94d853c0588222428e0c652
Bug-Url: https://bugzilla.redhat.com/1377682
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>