Accept 200 code for 'Add' methods
Currently the SDK doesn't accept the HTTP 200 code as a valid result of
'Add' methods, but some methods in the server return this code, for
example the method that adds a cluster quota limit. This patch changes
the SDK so that accepts it.

Change-Id: I88cb176b0ec53c77aac3edb11cecd24b2b93f9e8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 50a6347547f1c557caabe5c692a93fd6fce7e24d)