Convert 'all_content' parameter to 'All-Content' header
Older versions of the engine (before 4.1) only accepted the
'all_content' parameter as an HTTP header. To better support those
versions of the engine this patch changes the SDK so that when the
'all_content' parameter is specified its value is copied to the
'All-Content' request header. So a request like this:

  GET /ovirt-engine/api/vms/123?all_content=true

Will be translated into this:

  GET /ovirt-engine/api/vms/123?all_content=true
  All-Content: true

Change-Id: I3c15fa5cc60a2c45406d4d892c68d1b6f54d38e1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 9dedf6aeb19ef9dedc89ff0212103a8474be134b)