Run Rubocop as part of the 'test' task
This patch fixes most of the Rubocop complaints, and disables the cops
that can't be easily addressed, so that we can actually run Rubocop on
the source as part of the 'test' task.

Note that the generated code has also been improved to generate less
Rubocop complaints (1.9 hash syntax, no return, etc), but it isn't
actually checked because it takes too long, and some complaints (long
lines, for example) are hard to avoid.

Change-Id: Ia7268ae79cc55609b92c7cd17c71cfdcf9be9563
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>