Require Ruby 2
The SDK is no longer built or tested with Ruby 1.9, as it is now
unsupported and no longer available in the platforms supported by oVirt.
To make that explicit this patch changes the .gemspec file to indicate
that Ruby 2 is required.

Change-Id: Ia0f00ddae5acb26f053c72a3c3f12f8e2e776a5f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 38639f341c0c6690a61d07bb342907ee8cc1d75b)