Require json >= 1.8 instead of >= 2.0
Currently the SDK requires version 2 or newer of the 'json' gem, but
actually it doesn't need version 2, it can perfectly work with 1.8. In
addition Fedora and CentOS provide version 1.8, so the RPM generated for
the SDK won't work in those distributions. To solve this problem this
patch changes the SDK so that it requires version 1.8.

Change-Id: Ic1f5095c1f9181a91721484ab6838be78d4c0df4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>