Install built RPMs
In the past there have been situations where the built RPMs failed
during installation, because they required packages that don't exist in
the distribution. In order to detect that earlier in the future this
patch changes the automation scripts so that the built RPMs are
installed as part of the build process.

Change-Id: I34352228f6b394ce9ad82d5a80a45056550b3d47
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit aa52a827fca3afbd6da116214b80fb591ad91bdc)