Clarify the units of `provisioned_size`
This patch changes the example to add a VM disk so that it is clearer
that size of a disk has to be specified in bytes. It also changes the
example to use a more realistic size, 10 GiB, as it currently the size
is 1 MiB, which isn't realistic.

Change-Id: If21816ca1b180883bb799b32a2fd5857e7e25cd2
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 8e73bf7af2c9fea227ec8f9e4b84baebd5ea620f)