Use DiskAttachment.active explicitly
The 'DiskAttachment.active' attribute isn't mandatory when adding a disk
attachment to a virtual machine, but if it isn't provided the result is
undefined: the disk will be active in some situations, and won't in some
other situations. To avoid that uncertainty it is better to always
include it with the desired value. This patch changes the examples
accordingly.

Change-Id: Idc5675b5174aab5ec1707dd677c8bf8f1fe77483
Related-To: https://bugzilla.redhat.com/1390517
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 3c8488955185b383b6c537889e5cfb6d82c9d4b5)