Fix writing 'id' attribute for enum values
Currently the SDK doesn't generate correctly the XML for attributes
named 'id' that contain enum values. For example, when writing the list
of SSO methods of a virtual machine it generates the following:

  <sso>
    <methods>
      <method>
        <id>guest_agent</id>
      </method>
    </methods>
  </sso>

But it should write the following instead, as the 'id' model attribute
should always be represented as an XML attribute:

  <sso>
    <methods>
      <method id="guest_agent"/>
    </methods>
  </sso>

This patch addresses that issue, which was mostly a typo in the
generator of the XML writing code.

Change-Id: I211c87dba59c6c9a52858b8b0faa9ddd0cf99fda
Related-To: https://bugzilla.redhat.com/1408839
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit e77ad2991a1e11e5ffd5a99ce4700ba4897d20e3)