Enable installation in Vagrant with embedded Ruby
Currently when trying to install the SDK within as a Vagrant plugin the
build fails in some versions of Vagrant because the directories of the
embedded 'libxml2' and 'libcurl' libraries are incorrect. To avoid that
issue this patch changes the build scripts of the SDK so that they
explicitly fix those directories.

Change-Id: I8743b3b89161618e18e954d108f30aa649ba3edd
Bug-Url: https://bugzilla.redhat.com/1411594
Related-To: https://github.com/mitchellh/vagrant/issues/8184
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 338a7b246ebc7bfeb0cdcfc34d1870e2ea1ff877)