Accept CA files and strings
Currently the constructor of the 'Connection' class accepts the
'ca_file' option, which should contain the name of a file containing the
trusted CA certificates. This patch modifies it so that it also accepts
the 'ca_certs' option, containing an array of strings, each one
containing the PEM representation of one certificate.

Change-Id: I1f5ebc30147558fcd6093ba8f3a864a05517202a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>