Don't specify the 'json' version
Currently the SDK requires the 'json' gem using the following range
specification:

  spec.add_dependency('json', '~> 1.8')

This makes it incompatible with applications that want to use versio 2.0
or newer. But the SDK doesn't really need any advanced capability, it can
work with any version of the 'json' gem. To reduce the incompatibilities
this patch changes the SDK so that it doesn't specify the min version.
It will get 1.8 at least anyhow, has it is included with any Ruby 2
installation.

Change-Id: I62364b45e3782c17d51b3def0c4769e8e3aafed0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>