Use 'curl-config' to detect libcurl if possible
Currently the build process of the SDK expects the 'pkg-config' file of
the 'libcurl' library to be available, otherwise the build will fail. In
some environments (macOS, for example) this may not be available. But
curl also provides a 'curl-config' script, with similar capabilities,
that seems to be more widely available. This patch changes the
'extconf.rb' file, so that it first tries with this script, and, if it
fails, then tries with the 'pkg-config' mechanism.

Change-Id: I73e04663cfff34e7bc34d02d4a9b7493df4445b3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>