Add support for custom headers and query parameters
Currently there is no way to send to the server custom headers or query
parameters. This patch adds support for that. For example, if the
caller, for whatever the reason, needs to send the 'My: myvalue' header
and the 'your=yourvalue' query parameter to the request to add a virtual
machine, it can be done as follows:

  # Get the reference to the service that managse the collection
  # of virtual machines:
  vms_service = connection.system_service.vms_service

  # Add the virtual machine, with a custom
  vms_service.add(
    OvirtSDK4::Vm.new,
    headers: { my: 'myvalue' },
    query: { your: 'yourvalue' }
  )

Change-Id: I7f694bbff3889a62169eae894e2ba1071780237a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 598466b9aaa471d97bf2772db4d6234b0af85b9f)