Don't use 'bundler' in RPM distributions
Currently the SDK uses the 'bundler' tool to install the dependencies
and build the SDK. But in some platforms the latest version of this tool
doesn't work correctly with the version of the 'gem' tool. To avoid this
problem, and other potential problems in the future, this patch changes
the build scripts so that they provide a mechanism to build without
bundler. This mechanism consists on two new build profiles: 'bundler'
and 'rpm'. The first, 'bundler', will work as it used to work before
this patch. The new one, 'rpm', will assume that the required gems have
already been installed, with the RPM package manager, for example.

Note that 'bundler' is still the recommended tool for the build, and
will continue to be used in development environments and for the builds
released to 'rubygems.org'.

Change-Id: Ide97e6e6df3ca326f26563a7e3305fd50e0abc14
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 5b5d707aeedefe4745fd73c5b33735aa4f49f001)