Don't fork the Maven source plugin
Currently the project runs the 'jar' goal of the Maven source plugin.u
A side effect of this is that the plugin explicitly executes the
'generate-sources' phase of the build. As a result the code generation,
which is attached to that 'generate-sources' phase, runs twice. That
doesn't have a big impact, as the generated code is the same, but it
wastes resources. This patch changes the build process so that it uses
the 'jar-no-fork', which doesn't have that negative side effect.

Change-Id: I107a12125f5acb229608b05efba3572f24ebfe3d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit bdb11f9394b5d6e76e4cafeaf2e6aded0f2156e0)