Fix writing of lists of structs
Currently when the type of an attribute is a list of structs, the SDK
generates the XML using the name of the attribute as the tag for the
list, and the same name in singular as the tag for the elements. For
example, for the Bonding.slaves attribute it generates the following:

  <bonding>
    <slaves>
      <slave>
        <name>eth0</name>
      </slave>
      <slave>
        <name>eth1</name>
      </slave>
    </slaves>
  </bonding>

But the server expects the name of the type of the element:

  <bonding>
    <slaves>
      <host_nic>
        <name>eth0</name>
      </host_nic>
      <host_nic>
        <name>eth1</name>
      </host_nic>
    </slaves>
  </bonding>

Note the use of 'host_nic' instead of 'slave'.

The behaviour of the SDK is conceptually more correct than the behaviour
of the server, but as we need to support the existing server, without
changes, we need to fix the SDK instead.

Change-Id: I98722ac732e1fd6a5e2212138890639e6784e5d6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 3ba240f0c3f456d4ae81da9939fdbb314fa40eed)
(cherry picked from commit c5f3f87285ff5936707e2f6d129f828c3272921a)