Send INFO log messages with summaries of requests and responses
Currently the SDK only sends output the log when the 'debug' parameter
is 'true', and then it sends all the HTTP traffic. This is too much
information in most production environments, so an intermediate level of
information is required. This patch changes the SDK so that it will use
the 'INFO' level so send summaries of the requests sent and the
responses received, something like this:

  INFO: Sending 'GET' request to URL 'https://myserver/ovirt-engine/api/vms?search=name%3Dmyvm'.
  INFO: Received response code '200'.

This will be automatically enabled when passing a 'log' parameter to the
constructor of the 'Connection' class.

Change-Id: I38a81b8e646b8a327ccc3d80738e9d95c4055fc1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>