Check that URL has been provided
Currently the SDK doesn't check that the 'url' parameter has been
provided. As a unexpected errors happen later, when using the
connection, which are difficult to interpret. To avoid that this patch
changes the 'Connection' class so that it raises an exception is the
'url' parameter isn't provided.

Change-Id: I927c4097e718fe7b63e73c72bbff3c960b6387ac
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 772749b455a7601d4587e715d6a43f8a4cfbf73f)