Read correctly attributes with enum values
A recent patch introduced a mechanism to ignore incorrect values of enum
types. That patch also introduced a regresion for attributes of enum
types that have a name different to their type name. For example, the
'vm.migration.auto_converge' attribute is of type 'InheritableBoolean',
not of type 'AutoConverge' as assumed by the code generator. This patch
fixes that issue.

Change-Id: Ie0e67fd4ecf6e38b9e23c6b6cd929db1836adafd
Bug-Url: https://bugzilla.redhat.com/1428642
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>