Don't avoid reserved words in enum values text
The SDK appends an underscore to generated names that conflict with Ruby
reserved words. But it also does so for the text of enum values. This is
wrong, in particular it is wrong for the 'InheritableBoolean' enum which
has the following values currently:

  FALSE = 'false_'.freeze
  TRUE = 'true_'.freeze
  ...

The 'false_' and 'true_' values are incorrect, as that isn't what the
server generates or expects, they should be 'false' and 'true', without
the underscore.

This patch fixes that issue.

Change-Id: I5141e2be962d123934a35d14349d685f14e3bcd8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 4d3cb860592548e2805f530f8e9cb8343284ab4c)