Add support for asynchronous requests
This patch modifies the SDK so that it uses the 'multi' interface of the
'libcurl' library. This interface supports several advanced HTTP
capabilities, like multiple connections and pipelining.

The interface of the SDK is preserved, for backwards compatibility. The
only new addition is a new 'wait' parameter, for all operations. The
default value is 'true'. When the values is 'false' the operation is
performed asynchronously, and a 'future' object is returned that can be
used to retrieve the result later. For example:

  # Find the virtual machine and the service that manages it:
  vm = vms_service.list(search='name=myvm')[0]
  vm_service = vms_service.vm_service(vm.id)

  # Find the disk attachments, asynchronously:
  attachments_service = vm_service.disk_attachments_service
  attachments_future = attachments_service.list(wait: false)

  # Do some other tasks, for example sending some more asynchronous
  # requests, so that they are queued and potentially pipelined.

  # Later, wait for the result:
  attachments = attachments_future.wait

  # Print the details:
  attachments.each do |attachment|
    puts attachment.disk.id
  end

Change-Id: I8b004247ea79357eae768535fb3caf871329daf8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>