Automatically replace bad tokens
Currently when the token used by the SDK is no longer valid, because it
is expired or for any other reason, the following requests just fail.
The only way to recover is to close the connection and open a new one.
That isn't convenient for long running applications. To improve that
this patch modifies the SDK so that it automatically replaces bad
tokens, requesting new ones from the SSO service.

Change-Id: I5af5069dbd4ae3adbe9aa726053edc21bcbdfda7
Bug-Url: https://bugzilla.redhat.com/1434831
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit ae2554a954976739a51334409e6935f3f8fe9637)