Convert header names to lower case
This patch changes the SDK so that HTTP response header names are always
converted to lower case. That is needed in order to simplify the code
that handles headers, so that it can directly get them from the headers
hash instead of iterating it.

Change-Id: Id3ad00646a8c4d9906c80c29d0baf13c1b9a427e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 87da2121083c02783b672baf08e3c4b938c66289)