Add Error.fault
Currently when a fault is returned by the server the information that it
contains is used to generate an exception message, and then this
discarded. This patch changes the SDK so that the fault object is copied
to the 'Error.fault' attribute instead of discarded.

Change-Id: I98ff05baebb174bf9446b6098a28b3cb5da81021
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 37eb32c935c3e8e51d4013d7c221e613f99c159a)