Don't use curl_multi_wait if curl <= 7.28.0
The curl_multi_wait function was introduced in version 7.28.0 of
libcurl. That version of libcurl isn't available in some environments,
in particular it isn't available in the operating system used by default
in the Travis environment: Ubuntu Precise LTS. As a result the SDK fails
during runtime because of the missing symbol. To avoid that problem this
patch changes the SDK so that during build time it check the version of
the libcurl library, and if it is older than 7.28.0 it uses the 'select'
system call instead.

Change-Id: I2e15d50601baf35445a5162abefd1d67ed24542c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>