Don't run the tests if -DskipTests=true
Currently the tests of the SDK are executed even when the
'-DskipTests=true' option is given in the Maven command line. A side
effect of this is that the tests run twice in Travis. To avoid that this
patch changes the POM of the SDK so that the tests will not be executed
in this scenario.

Change-Id: Iecdca3b98478279a3923c0382ad7494017526b41
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 43814c08daf338bc078b3aafe3b015dddf52de06)