Ignore unrelated responses from server
When using for example Kerberos authentication the server can return to
the client more then one response. The client is interested only in last
response which contains the information about the authentication, if it
was successfull, so this patch add logic to ignore the headers which
don't belong to last response.

Change-Id: I3d894361dc15f6c21941db254c567d2d5ab9fb6f
Related-to: https://bugzilla.redhat.com/1459254
Signed-off-by: Ondra Machacek <omachace@redhat.com>
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit af184796a6c15e7f3b3148c7edf7d053c5fe6142)